RDMA/nldev: add provider-specific resource tracking
Each provider can register a "fill entry" function with the restrack core.
This function will be called when filling out a resource, allowing the
provider to add provider-specific details.  The details consist of a
table of nested attributes, that are in the form of "key, value" tuple.
The key nlattr must be strings, and the value nlattr can be one of
provider attributes that are generic, but typed, allowing the nlmessage
to ve validated.  Currently the types include string, s32, u32, x32, s64,
u64, and x64. The inclusion of x, s, and u variants for numeric attributes
allows the user tool to print the number in the desired format.

More attrs can be defined as the become needed by providers.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>