RDMA/IWPM: Support no port mapping requirements
A soft iwarp driver that uses the host TCP stack via a kernel mode
socket does not need port mapping.  In fact, if the port map daemon,
iwpmd, is running, then iwpmd must not try and create/bind a socket to
the actual port for a soft iwarp connection, since the driver already
has that socket bound.

Yet if the soft iwarp driver wants to interoperate with hard iwarp
devices that -are- using port mapping, then the soft iwarp driver's
mappings still need to be maintained and advertised by the iwpm protocol.

This patch enhances the rdma driver<->iwcm interface to allow an
iwarp driver to specify that it does not want port mapping.  The iwpm
kernel<->iwpmd interface is also enhanced to pass up this information
on map requests.

Care is taken (hopefully) to interoperate with the current iwpmd version
(ABI version 3) and only use the new NL attributes if iwpmd supports
ABI version 4.

In order to add the new IWPM_NL attribute, the enum for the IWPM map
command attributes are cleaned up such that the new attribute can be
added without breaking ABI version 3: Instead of mixing the request
and response attributes, separate enums are created and used.  They
are created, though, in a way to avoid breaking ABI version 3.

The ABI define has also been created in rdma_netlink.h so both kernel
and user code can share it.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>