lib/irq_poll: support schedules in non-interrupt contexts
Do not assume irq_poll_sched() is called from an interrupt context only.
So use raise_softirq_irqoff() instead off __raise_softirq_irqoff() so it
will kick the ksoftirqd if the schedule is from a non-interrupt context.

This is required for RDMA drivers, like soft iwarp, that generate cq
completion notifications in a workqueue or kthread context.  Without this
change, siw completion notifications to the ULP can take several hundred
usecs, depending on the system load.

Signed-off-by: Steve Wise <swise@opengridcomputing.com>